+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [October 18, 2024]
+
+### Added
+- CHANGELOG
+- Add analysis.py for annotation
+- Modify the embedding and add plot_expr to enable gene expression across batchßß
+
 ## [April 29, 2024]
 
 ### Added